refactor: move server role components into own folder
