+++ src/components/Roles/ServerRoles/EditServerRoleTabs.tsx
@@ -2,10 +2,10 @@ import { Card, Tab, Tabs } from "@mui/material";
 import { useRouter } from "next/router";
 import { SyntheticEvent, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { EditServerRoleTabsFragment } from "../../apollo/gen";
-import { EditRoleTabNames } from "../../constants/role.constants";
-import { useAboveBreakpoint } from "../../hooks/common.hooks";
-import AddRoleMemberTab from "./AddRoleMemberTab";
+import { EditServerRoleTabsFragment } from "../../../apollo/gen";
+import { EditRoleTabNames } from "../../../constants/role.constants";
+import { useAboveBreakpoint } from "../../../hooks/common.hooks";
+import AddRoleMemberTab from "../AddRoleMemberTab";
 import DeleteServerRoleButton from "./DeleteServerRoleButton";
 import ServerPermissionsForm from "./ServerPermissionsForm";
 import ServerRoleForm from "./ServerRoleForm";
