+++ src/pages/roles/index.tsx
@@ -2,7 +2,7 @@ import { Typography } from "@mui/material";
 import { NextPage } from "next";
 import { useTranslation } from "react-i18next";
 import { useServerRolesQuery } from "../../apollo/gen";
-import ServerRoleForm from "../../components/Roles/ServerRoleForm";
+import ServerRoleForm from "../../components/Roles/ServerRoles/ServerRoleForm";
 import RoleList from "../../components/Roles/RoleList";
 import LevelOneHeading from "../../components/Shared/LevelOneHeading";
 import ProgressBar from "../../components/Shared/ProgressBar";
